Added Scenes and updating!
Scenes are now containers for a single world and props to put in it.
So far, there is not much flexibility, you just create a scene and put
your props in it.
Then you can use createWorld() to create a minecraft world with your
props in it.
Then you can use addPlayer() to teleport a player to the scene.
And there is also a thread that is calling doUpdateTick() every so
often.
It calls update() on all of its props, passing in how much time since
the last tick.
So now things can be animated and cool stuff like that!

Also, GridWorldGenerator is part of engine, not editor, now.
Also, EditorWorld is now a subclass of Scene.