Fixed editor bugs.
Added custom implementations of getPosition() and getRealPosition() in
EntityProp so that if the entity moves by itself, the changes in
position are reflected in the code.
Modified how position is modified for props to make ^^^ work.
Made the way toolbars are built much more flexible by putting
sub-toolbars in separate methods and by removing the position argument
from addItem().