Added Animations! (Sort of...)
Added AnimationData, which holds keyframes and interpolates between
them.
Added AnimatableProperty, which holds one or more channels of
AnimationData to animate different data types. Subclasses of this should
eventually replace regular member variables in props.
Added AnimatableDouble and AnimatableVector, subclasses of
AnimatableProperty for animating Doubles and Vectors.
Added AnimationListener, an interface that is notified when an animation
changes a value.
Added Animation, which holds a particular animation (walking, running,
swimming, etc.)
Added AnimPropertySet, which holds several AnimationProperties and
Animations used for them.

So far, only a single, static, 1-frame animation for prop location is
implemented, but the framework is in place. The next step is replacing
member variables with AnimatableProperties.